Navigation changed

+++ app/src/main/java/com/example/journalchat/data/repositories/ChatRepository.kt
@@ -1,7 +1,10 @@
 package com.example.journalchat.data.repositories
 
+import androidx.room.Query
 import com.example.journalchat.data.daos.ChatDao
 import com.example.journalchat.data.models.Chat
+import com.example.journalchat.data.models.Message
+import com.example.journalchat.ui.uiModels.ChatUi
 import kotlinx.coroutines.flow.Flow
 
 
@@ -9,11 +12,14 @@ class ChatRepository(private val chatDao: ChatDao) {
 
     fun getAllStream(): Flow<List<Chat>> = chatDao.getAll()
 
-    fun getStream(id: Int): Flow<Chat?> = chatDao.get(id)
+    fun getStream(id: Long): Flow<Chat?> = chatDao.get(id)
 
     suspend fun insertItem(item: Chat) = chatDao.insert(item)
 
     suspend fun deleteItem(item: Chat) = chatDao.delete(item)
 
     suspend fun updateItem(item: Chat) = chatDao.update(item)
+
 }
+
+
